feat(reader): unsubscribe from topic
