+++ src/lib/omerlo/reader/endpoints/device.ts
@@ -1,0 +1,20 @@
+import { request } from '$reader/utils/request';
+
+export const deviceFetchers = (f: typeof fetch) => {
+  return {
+    registerDevice: registerDevice(f),
+  };
+};
+
+export interface DeviceParams {
+  pushToken: string,
+  name: string,
+}
+
+export function registerDevice(f: typeof fetch) {
+  return (params: DeviceParams) => {
+    const body = { push_token: params.pushToken, name: params.name };
+    const opts = { body, method: 'post' };
+    return request(f, '/devices/register', opts);
+  };
+}
